pyproject.toml
[project]
name = "eqm_mnist"
version = "0.1.0"
description = ""
authors = [
    {name = "Luke Martin",email = "lukechile@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib (>=3.10.3,<4.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "ffmpeg (>=1.4,<2.0)",
    "celluloid (>=0.2.0,<0.3.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
